Move requests to model
