+++ services/skus/model/model.go
@@ -286,3 +286,15 @@ func (x *OrderTimeBounds) ExpiresAtWithFallback(fallback time.Time) time.Time {
 
 	return expiresAt
 }
+
+// CreateOrderRequest includes information needed to create an order.
+type CreateOrderRequest struct {
+	Email string             `json:"email" valid:"-"`
+	Items []OrderItemRequest `json:"items" valid:"-"`
+}
+
+// OrderItemRequest represents an item in a order request.
+type OrderItemRequest struct {
+	SKU      string `json:"sku" valid:"-"`
+	Quantity int    `json:"quantity" valid:"int"`
+}
